Change enum class base from size_t to unsigned int
I have checked on GCC and clang...

On 64-bit CPUs, size_t is typically 64 bits but the default for enum
class appears to be 32 bits so it makes more sense to use unsigned int
as the base type when it has to be specified (for forward declarations).

+++ source/scan/detail/check_scan/test_only/check_scan_on_final_alignment.h
@@ -35,7 +35,7 @@ namespace cath { namespace scan { class scan_stride; } }
 namespace cath { namespace scan { namespace detail { class alignment_scan_comparison; } } }
 namespace cath { namespace scan { namespace detail { class multi_struc_res_rep_pair; } } }
 namespace cath { namespace scan { namespace detail { class quad_and_rep_criteria_result; } } }
-namespace cath { namespace scan { namespace detail { enum class quad_criteria_result : size_t; } } }
+namespace cath { namespace scan { namespace detail { enum class quad_criteria_result : unsigned int; } } }
 namespace cath { namespace scan { namespace detail { class roled_scan_stride; } } }
 namespace cath { namespace scan { namespace detail { class single_struc_res_pair; } } }
